+++ WebSocket.js
@@ -8,4 +8,3 @@ export function websocket(uri, on_message, on_open=empty,  on_close=empty) {
 	ws.onclose = on_close;
 	return ws;
 }
-